Make it explicit that assigning points to m.VALM is a cast to uint8_t. Add a comment about using ML elements rather than just pointers in the PLYIX[] table. Still intend making the PLYIX[] table code clearer

+++ src/zargon.cpp
@@ -1678,7 +1678,7 @@ void POINTS()
 
     // Save score value
     // Experiment: after rescale -126->2 ... -1->127, 0->128, 1->129 ... 126->254
-    m.VALM = points;
+    m.VALM = (uint8_t)points;
 
     // Save score value to move pointed to by current move ptr
     m.MLPTRJ->val = m.VALM;
